+++ returns/context/requires_context_future_result.py
@@ -18,10 +18,12 @@ from returns._generated.futures import _future_result, _reader_future_result
 from returns._generated.iterable import iterable
 from returns.context import NoDeps
 from returns.future import Future, FutureResult
+from returns.hkt import Kind, dekind
 from returns.io import IO, IOResult
 from returns.primitives.container import BaseContainer
 from returns.primitives.types import Immutable
 from returns.result import Result
+from returns.typeclasses import applicative, functor, monad
 
 if TYPE_CHECKING:
     from returns.context.requires_context import RequiresContext
@@ -46,7 +48,10 @@ _FirstType = TypeVar('_FirstType')
 @final
 class RequiresContextFutureResult(
     BaseContainer,
-    Generic[_EnvType, _ValueType, _ErrorType],
+    Kind['RequiresContextFutureResult', _ValueType, _ErrorType, _EnvType],
+    functor.Functor[_ValueType],
+    applicative.Applicative[_ValueType],
+    monad.Monad[_ValueType],
 ):
     """
     The ``RequiresContextFutureResult`` combinator.
@@ -164,7 +169,7 @@ class RequiresContextFutureResult(
     def map(  # noqa: WPS125
         self,
         function: Callable[[_ValueType], _NewValueType],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Composes successful container with a pure function.
 
@@ -189,9 +194,13 @@ class RequiresContextFutureResult(
 
     def apply(
         self,
-        container: 'RequiresContextFutureResult['
-            '_EnvType, Callable[[_ValueType], _NewValueType], _ErrorType]',
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+        container: Kind[
+            'RequiresContextFutureResult',
+            Callable[[_ValueType], _NewValueType],
+            _ErrorType,
+            _EnvType,
+        ],
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Calls a wrapped function in a container on this container.
 
@@ -220,16 +229,21 @@ class RequiresContextFutureResult(
 
         """
         return RequiresContextFutureResult(
-            lambda deps: self(deps).apply(container(deps)),
+            lambda deps: self(deps).apply(dekind(container)(deps)),
         )
 
     def bind(
         self,
         function: Callable[
             [_ValueType],
-            'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]',
+            Kind[
+                'RequiresContextFutureResult',
+                _NewValueType,
+                _ErrorType,
+                _EnvType,
+            ],
         ],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Composes this container with a function returning the same type.
 
@@ -260,7 +274,7 @@ class RequiresContextFutureResult(
         """
         return RequiresContextFutureResult(
             lambda deps: self(deps).bind(
-                lambda inner: function(inner)(deps),  # type: ignore[misc]
+                lambda inner: dekind(function(inner))(deps),  # type: ignore[misc]
             ),
         )
 
@@ -270,10 +284,10 @@ class RequiresContextFutureResult(
             [_ValueType],
             Awaitable[
                 'RequiresContextFutureResult'
-                '[_EnvType, _NewValueType, _ErrorType]'
+                '[_NewValueType, _ErrorType, _EnvType]'
             ],
         ],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Composes this container with a async function returning the same type.
 
@@ -311,7 +325,7 @@ class RequiresContextFutureResult(
     def bind_awaitable(
         self,
         function: Callable[[_ValueType], 'Awaitable[_NewValueType]'],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Allows to compose a container and a regular ``async`` function.
 
@@ -350,7 +364,7 @@ class RequiresContextFutureResult(
     def bind_result(
         self,
         function: Callable[[_ValueType], 'Result[_NewValueType, _ErrorType]'],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``Result`` returning function to the current container.
 
@@ -387,9 +401,9 @@ class RequiresContextFutureResult(
         self,
         function: Callable[
             [_ValueType],
-            'RequiresContext[_EnvType, _NewValueType]',
+            'RequiresContext[_NewValueType, _EnvType]',
         ],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``RequiresContext`` returning function to current container.
 
@@ -427,9 +441,9 @@ class RequiresContextFutureResult(
         self,
         function: Callable[
             [_ValueType],
-            'RequiresContextResult[_EnvType, _NewValueType, _ErrorType]',
+            'RequiresContextResult[_NewValueType, _ErrorType, _EnvType]',
         ],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``RequiresContextResult`` returning function to the current one.
 
@@ -470,9 +484,9 @@ class RequiresContextFutureResult(
         self,
         function: Callable[
             [_ValueType],
-            'RequiresContextIOResult[_EnvType, _NewValueType, _ErrorType]',
+            'RequiresContextIOResult[_NewValueType, _ErrorType, _EnvType]',
         ],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``RequiresContextIOResult`` returning function to the current one.
 
@@ -511,7 +525,7 @@ class RequiresContextFutureResult(
     def bind_io(
         self,
         function: Callable[[_ValueType], IO[_NewValueType]],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``IO`` returning function to the current container.
 
@@ -542,7 +556,7 @@ class RequiresContextFutureResult(
     def bind_ioresult(
         self,
         function: Callable[[_ValueType], IOResult[_NewValueType, _ErrorType]],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``IOResult`` returning function to the current container.
 
@@ -577,7 +591,7 @@ class RequiresContextFutureResult(
     def bind_future(
         self,
         function: Callable[[_ValueType], Future[_NewValueType]],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``Future`` returning function to the current container.
 
@@ -615,7 +629,7 @@ class RequiresContextFutureResult(
             [_ValueType],
             FutureResult[_NewValueType, _ErrorType],
         ],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``FutureResult`` returning function to the current container.
 
@@ -650,7 +664,7 @@ class RequiresContextFutureResult(
     def bind_async_future(
         self,
         function: Callable[[_ValueType], Awaitable[Future[_NewValueType]]],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Binds ``Future`` returning async function to the current container.
 
@@ -688,7 +702,7 @@ class RequiresContextFutureResult(
             [_ValueType],
             Awaitable[FutureResult[_NewValueType, _ErrorType]],
         ],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Bind ``FutureResult`` returning async function to the current container.
 
@@ -724,7 +738,7 @@ class RequiresContextFutureResult(
 
     def fix(
         self, function: Callable[[_ErrorType], _NewValueType],
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, _ErrorType, _EnvType]':
         """
         Composes failed container with a pure function.
 
@@ -755,7 +769,7 @@ class RequiresContextFutureResult(
 
     def alt(
         self, function: Callable[[_ErrorType], _NewErrorType],
-    ) -> 'RequiresContextFutureResult[_EnvType, _ValueType, _NewErrorType]':
+    ) -> 'RequiresContextFutureResult[_ValueType, _NewErrorType, _EnvType]':
         """
         Composes failed container with a pure function.
 
@@ -788,9 +802,9 @@ class RequiresContextFutureResult(
         self,
         function: Callable[
             [_ErrorType],
-            'RequiresContextFutureResult[_EnvType, _ValueType, _NewErrorType]',
+            'RequiresContextFutureResult[_ValueType, _NewErrorType, _EnvType]',
         ],
-    ) -> 'RequiresContextFutureResult[_EnvType, _ValueType, _NewErrorType]':
+    ) -> 'RequiresContextFutureResult[_ValueType, _NewErrorType, _EnvType]':
         """
         Composes this container with a function returning the same type.
 
@@ -914,7 +928,7 @@ class RequiresContextFutureResult(
     @classmethod
     def from_result(
         cls, inner_value: Result[_ValueType, _ErrorType],
-    ) -> 'RequiresContextFutureResult[NoDeps, _ValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_ValueType, _ErrorType, NoDeps]':
         """
         Creates new container with ``Result`` as a unit value.
 
@@ -944,7 +958,7 @@ class RequiresContextFutureResult(
     def from_io(
         cls,
         inner_value: IO[_NewValueType],
-    ) -> 'RequiresContextFutureResult[NoDeps, _NewValueType, Any]':
+    ) -> 'RequiresContextFutureResult[_NewValueType, Any, NoDeps]':
         """
         Creates new container from successful ``IO`` value.
 
@@ -968,7 +982,7 @@ class RequiresContextFutureResult(
     def from_failed_io(
         cls,
         inner_value: IO[_NewErrorType],
-    ) -> 'RequiresContextFutureResult[NoDeps, Any, _NewErrorType]':
+    ) -> 'RequiresContextFutureResult[Any, _NewErrorType, NoDeps]':
         """
         Creates a new container from failed ``IO`` value.
 
@@ -991,7 +1005,7 @@ class RequiresContextFutureResult(
     @classmethod
     def from_ioresult(
         cls, inner_value: IOResult[_ValueType, _ErrorType],
-    ) -> 'RequiresContextFutureResult[NoDeps, _ValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_ValueType, _ErrorType, NoDeps]':
         """
         Creates new container with ``IOResult`` as a unit value.
 
@@ -1020,7 +1034,7 @@ class RequiresContextFutureResult(
     def from_future(
         cls,
         inner_value: Future[_ValueType],
-    ) -> 'RequiresContextFutureResult[NoDeps, _ValueType, Any]':
+    ) -> 'RequiresContextFutureResult[_ValueType, Any, NoDeps]':
         """
         Creates new container with successful ``Future`` as a unit value.
 
@@ -1045,7 +1059,7 @@ class RequiresContextFutureResult(
     def from_failed_future(
         cls,
         inner_value: Future[_ErrorType],
-    ) -> 'RequiresContextFutureResult[NoDeps, Any, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[Any, _ErrorType, NoDeps]':
         """
         Creates new container with failed ``Future`` as a unit value.
 
@@ -1072,7 +1086,7 @@ class RequiresContextFutureResult(
     def from_future_result(
         cls,
         inner_value: FutureResult[_ValueType, _ErrorType],
-    ) -> 'RequiresContextFutureResult[NoDeps, _ValueType, _ErrorType]':
+    ) -> 'RequiresContextFutureResult[_ValueType, _ErrorType, NoDeps]':
         """
         Creates new container with ``FutureResult`` as a unit value.
 
@@ -1104,8 +1118,8 @@ class RequiresContextFutureResult(
     def from_typecast(
         cls,
         inner_value: 'RequiresContext['
-            '_EnvType, FutureResult[_NewValueType, _NewErrorType]]',
-    ) -> 'RequiresContextFutureResult[_EnvType, _NewValueType, _NewErrorType]':
+            'FutureResult[_NewValueType, _NewErrorType], _EnvType]',
+    ) -> 'RequiresContextFutureResult[_NewValueType, _NewErrorType, _EnvType]':
         """
         You might end up with ``RequiresContext[FutureResult]`` as a value.
 
@@ -1140,8 +1154,8 @@ class RequiresContextFutureResult(
 
     @classmethod
     def from_context(
-        cls, inner_value: 'RequiresContext[_EnvType, _FirstType]',
-    ) -> 'RequiresContextFutureResult[_EnvType, _FirstType, Any]':
+        cls, inner_value: 'RequiresContext[_FirstType, _EnvType]',
+    ) -> 'RequiresContextFutureResult[_FirstType, Any, _EnvType]':
         """
         Creates new container from ``RequiresContext`` as a success unit.
 
@@ -1165,8 +1179,8 @@ class RequiresContextFutureResult(
 
     @classmethod
     def from_failed_context(
-        cls, inner_value: 'RequiresContext[_EnvType, _FirstType]',
-    ) -> 'RequiresContextFutureResult[_EnvType, Any, _FirstType]':
+        cls, inner_value: 'RequiresContext[_FirstType, _EnvType]',
+    ) -> 'RequiresContextFutureResult[Any, _FirstType, _EnvType]':
         """
         Creates new container from ``RequiresContext`` as a failure unit.
 
@@ -1191,8 +1205,8 @@ class RequiresContextFutureResult(
     @classmethod
     def from_result_context(
         cls,
-        inner_value: 'RequiresContextResult[_EnvType, _ValueType, _ErrorType]',
-    ) -> 'RequiresContextFutureResult[_EnvType, _ValueType, _ErrorType]':
+        inner_value: 'RequiresContextResult[_ValueType, _ErrorType, _EnvType]',
+    ) -> 'RequiresContextFutureResult[_ValueType, _ErrorType, _EnvType]':
         """
         Creates new container from ``RequiresContextResult`` as a unit value.
 
@@ -1225,8 +1239,8 @@ class RequiresContextFutureResult(
     def from_ioresult_context(
         cls,
         inner_value:
-            'RequiresContextIOResult[_EnvType, _ValueType, _ErrorType]',
-    ) -> 'RequiresContextFutureResult[_EnvType, _ValueType, _ErrorType]':
+            'RequiresContextIOResult[_ValueType, _ErrorType, _EnvType]',
+    ) -> 'RequiresContextFutureResult[_ValueType, _ErrorType, _EnvType]':
         """
         Creates new container from ``RequiresContextIOResult`` as a unit value.
 
@@ -1258,7 +1272,7 @@ class RequiresContextFutureResult(
     @classmethod
     def from_value(
         cls, inner_value: _FirstType,
-    ) -> 'RequiresContextFutureResult[NoDeps, _FirstType, Any]':
+    ) -> 'RequiresContextFutureResult[_FirstType, Any, NoDeps]':
         """
         Creates new container with successful ``FutureResult`` as a unit value.
 
@@ -1280,7 +1294,7 @@ class RequiresContextFutureResult(
     @classmethod
     def from_failure(
         cls, inner_value: _FirstType,
-    ) -> 'RequiresContextFutureResult[NoDeps, Any, _FirstType]':
+    ) -> 'RequiresContextFutureResult[Any, _FirstType, NoDeps]':
         """
         Creates new container with failed ``FutureResult`` as a unit value.
 
@@ -1304,9 +1318,9 @@ class RequiresContextFutureResult(
         cls,
         inner_value:
             Iterable[
-                'RequiresContextFutureResult[_EnvType, _ValueType, _ErrorType]',
+                'RequiresContextFutureResult[_ValueType, _ErrorType, _EnvType]',
             ],
-    ) -> 'ReaderFutureResult[_EnvType, Sequence[_ValueType], _ErrorType]':
+    ) -> 'ReaderFutureResult[Sequence[_ValueType], _ErrorType, _EnvType]':
         """
         Transforms an iterable of ``RequiresContextFutureResult`` containers.
 
@@ -1351,7 +1365,7 @@ class ContextFutureResult(Immutable, Generic[_EnvType], metaclass=ABCMeta):
     __slots__ = ()
 
     @classmethod
-    def ask(cls) -> RequiresContextFutureResult[_EnvType, _EnvType, Any]:
+    def ask(cls) -> RequiresContextFutureResult[_EnvType, Any, _EnvType]:
         """
         Is used to get the current dependencies inside the call stack.
 
@@ -1383,13 +1397,15 @@ class ContextFutureResult(Immutable, Generic[_EnvType], metaclass=ABCMeta):
 
 #: Alias for a popular case when ``Result`` has ``Exception`` as error type.
 RequiresContextFutureResultE = RequiresContextFutureResult[
-    _EnvType, _ValueType, Exception,
+    _ValueType, Exception, _EnvType,
 ]
 
 #: Sometimes `RequiresContextFutureResult` is too long to type.
-ReaderFutureResult = RequiresContextFutureResult
+ReaderFutureResult = RequiresContextFutureResult[
+    _ValueType, _ErrorType, _EnvType,
+]
 
 #: Alias to save you some typing. Uses ``Exception`` as error type.
 ReaderFutureResultE = RequiresContextFutureResult[
-    _EnvType, _ValueType, Exception,
+    _ValueType, Exception, _EnvType,
 ]