+++ CHANGELOG.md
@@ -21,6 +21,7 @@ See [0Ver](https://0ver.org/).
 - Adds `dekind` function to turn `Kind[IO, T]` into `IO[T]`
 - Adds new typeclasses for all public interfaces
 - Adds `collect_trace` helper function for better development experience
+- Adds missing `from_requires_context_future_result` to `RequiresContext`
 
 
 ## 0.14.0